Move createBitmapFromEncodedString() to ImageUtils
