+++ app/src/main/java/org/kiwix/kiwixmobile/history/HistoryActivity.java
@@ -30,7 +30,7 @@ import javax.inject.Inject;
 import butterknife.BindView;
 
 import static org.kiwix.kiwixmobile.utils.Constants.EXTRA_CHOSE_X_URL;
-import static org.kiwix.kiwixmobile.zim_manager.library.LibraryAdapter.createBitmapFromEncodedString;
+import static org.kiwix.kiwixmobile.utils.ImageUtils.createBitmapFromEncodedString;
 
 public class HistoryActivity extends BaseActivity implements HistoryContract.View,
     HistoryAdapter.OnItemClickListener {
@@ -230,7 +230,7 @@ public class HistoryActivity extends BaseActivity implements HistoryContract.Vie
 
   private void toggleSelection(ImageView favicon, History history) {
     if (deleteList.remove(history)) {
-      favicon.setImageBitmap(createBitmapFromEncodedString(history.getFavicon(), this));
+      favicon.setImageBitmap(createBitmapFromEncodedString(this, history.getFavicon()));
     } else {
       favicon.setImageDrawable(ContextCompat.getDrawable(this, R.drawable.ic_check_circle_blue_24dp));
       deleteList.add(history);
